02_Projects/woosdom_app/requirements.txt
flask>=3.0
pywebview>=5.0
jinja2>=3.1
py2app
